interface/app/requirements.txt
fastapi[standard]>=0.115.8
pyserial>=3.5
